+++ packages/vue/src/toggle/index.ts
@@ -1,0 +1,8 @@
+import { Toggle as ToggleRoot, type ToggleProps } from './toggle'
+
+const Toggle = Object.assign(ToggleRoot, {
+  Root: ToggleRoot,
+})
+
+export { Toggle }
+export type { ToggleProps }