+++ packages/react/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Added `Toggle` component
+
 ### Fixed
 
 - Fix issue where event callbacks that use `flushSync` did not have a stable reference, resulting in a noticable blocking re-renders.