+++ packages/vue/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Added `Toggle` component
+
 ### Changed
 
 - Revised `Pressable` component